+++ csatar-plugins/knowledgerepository/lang/hu/lang.php
@@ -20,6 +20,7 @@
                     'headCount' => 'Létszám',
                     'duration' => 'Időtartam',
                     'location' => 'Helyszín',
+                    'gameType' => 'Játék típus',
                 ],
             ],
             'general' => [