+++ Changelog.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- CS-384 - Created GameType model, controller, seeder, backend list and form
 - CS-371 - Created Location model, controller, seeder, backend list and form
 - CS-369 - Created Duration model, controller, seeder, backend list and form
 - CS-375 - Created Tool model, controller, seeder backend list and form